Server: Split conversion logic
