+++ Server/Converters/Currency.cs
@@ -1,0 +1,11 @@
+﻿namespace Server.Converters
+{
+    public struct Currency
+    {
+        public string AmountSingular { get; init; }
+        public string AmountPlural { get; init; }
+        public string CoinsSingular { get; init; }
+        public string CoinsPlural { get; init; }
+        public string Symbol { get; init; }
+    }
+}
